Post process script creates same folder structure as required by GTAModel Updated output folder and file names Updated SQL table column types to be more memory efficient

+++ scripts/GTAModelTransform.sql
@@ -5,7 +5,7 @@ drop table if exists GTA_PERSONS;
 
 CREATE TABLE GTA_PERSONS(HouseholdId INT(11),
 PersonNumber INT(11),
-Age Int,
+Age smallint,
 Sex VARCHAR(1),
 License VARCHAR(1),
 TransitPass VARCHAR(1) DEFAULT '0',
@@ -13,9 +13,9 @@ EmploymentStatus VARCHAR(1),
 Occupation VARCHAR(1),
 FreeParking VARCHAR(1) DEFAULT '0',
 StudentStatus VARCHAR(1),
-EmploymentZone INT DEFAULT 0,
-SchoolZone INT DEFAULT  0,
-ExpansionFactor INT )
+EmploymentZone smallint DEFAULT 0,
+SchoolZone smallint DEFAULT  0,
+ExpansionFactor smallint )
 as SELECT tempId as HouseholdId, PersonNumber, Age, Sex, License, '0' as TransitPass,
           EmploymentStatus, Occupation, StudentStatus, '0' as FreeParking,  EmploymentZone,
           '0' as SchoolZone, finalweight as ExpansionFactor
@@ -26,18 +26,18 @@ drop table if exists GTA_HOUSEHOLDS;
 #HouseholdId,HouseholdZone,ExpansionFactor,NumberOfPersons,DwellingType,Vehicles,IncomeClass
 CREATE TABLE GTA_HOUSEHOLDS(
 HouseholdId INT(11),
-HouseholdZone INT(11),
-ExpansionFactor INT (11),
-NumberOfPersons INT(11),
-DwellingType INT(11),
-Vehicles INT(11),
-IncomeClass INT(11)) AS SELECT tempId as HouseholdId, maz as HouseholdZone, finalweight as ExpansionFactor, DwellingType,NumberOfPersons,Vehicles,IncomeClass from synpop_hh
+HouseholdZone SMALLINT,
+ExpansionFactor SMALLINT,
+NumberOfPersons TINYINT,
+DwellingType TINYINT,
+Vehicles TINYINT,
+IncomeClass TINYINT) AS SELECT tempId as HouseholdId, maz as HouseholdZone, finalweight as ExpansionFactor, DwellingType,NumberOfPersons,Vehicles,IncomeClass from synpop_hh
 order by tempId asc;
 
 drop table if exists zonal_totals;
 
 CREATE TABLE ZONAL_TOTALS(
-    Zone INT NOT NULL DEFAULT 0,
+    Zone SMALLINT NOT NULL DEFAULT 0,
     ExpandedHouseholds INT NOT NULL DEFAULT 0
 )
 AS (SELECT  HouseholdZone as Zone, SUM(ExpansionFactor) as ExpandedHouseholds from GTA_HOUSEHOLDS group by HouseholdZone)
